+++ electron-builder.yml
@@ -10,6 +10,9 @@ mac:
       - target: "dmg"
         arch: 
           - "x64"
+      - target: "zip"
+        arch: 
+          - "x64"
 files:
   - "app"
   - "html"
