+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "radix-onboard",
-  "version": "1.0.5",
+  "version": "1.0.8",
   "description": "",
   "main": "main.js",
   "scripts": {
